+++ src/CoilpackServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace Expressionengine\Coilpack;
 
+use Expressionengine\Coilpack\Auth\CoilpackUserProvider;
 use Expressionengine\Coilpack\Api\Graph\Support\FieldtypeRegistrar;
 use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Event;